Made finite state machine more readable. Also refined find_best_edge coordinates2 function. Additionally, added function to got from image place coordinates to camera coordinate using depth as input. Lastly, depth is now measured in cm when considerring the thresholds.

+++ sw/airborne/modules/wedgebug/wedgebug.h
@@ -66,8 +66,8 @@ extern int SE_dilation_OCV_2; // SE size for the second dilation operation (see 
 
 // Setting thresholds
 
-extern float threshold_median_depth;		//! Below this median depth (m), an obstacle is considered to block the way (i.e. the blocking obstacle needs to be close)
-extern float threshold_depth_of_edge;	//! Below this depth (m) edges are eligible for the WedgeBug algorith
+extern uint16_t threshold_median_depth;		//! Below this median depth (m), an obstacle is considered to block the way (i.e. the blocking obstacle needs to be close)
+extern uint16_t threshold_depth_of_edge;	//! Below this depth (m) edges are eligible for the WedgeBug algorith
 
 extern int threshold_edge_magnitude;  			// Edges with a magnitude above this value are detected. Above this value, edges are given the value 127, otherwise they are given the value zero.
 extern float threshold_distance_to_goal; 		// Above this threshold, the goal is considered reached
